requirements.txt
flask==2.3.3
langchain==0.1.20
langchain-core==0.1.53
langchain-community==0.0.38
langchain-groq==0.1.3
sentence-transformers==2.2.2
transformers==4.44.2
huggingface-hub==0.24.6
faiss-cpu==1.7.4
numpy==1.23.5
Pillow==9.5.0
scikit-learn==1.2.2
tqdm==4.65.0
pydantic==1.10.12
tensorflow==2.12.0
keras==2.12.0
tensorboard==2.12.3
protobuf==4.25.8
gunicorn==21.2.0
flask-cors==3.0.10